feat: added rbac rules and permissions

+++ src/controllers/web/DefaultController.php
@@ -20,17 +20,16 @@ class DefaultController extends Controller
     
     public function actionIndex()
     {
+        if (!\Yii::$app->user->can('storageWebDefaultIndex') && !\Yii::$app->user->can('storageWebDefaultIndexOwn')) {
+            throw new \yii\web\ForbiddenHttpException(Module::t('You are not allowed to access this page.'));
+        }
         $model = new Storage();
         $searchModel = new StorageSearch();
         $id_directory = Yii::$app->request->get('id_directory');
         $dataProvider = $searchModel->search($this->request->queryParams);
-        
-        // File data provider setup
         $fileDataProvider = $searchModel->search(Yii::$app->request->queryParams);
         $fileDataProvider->query->andWhere(['id_directory' => $id_directory]);
         $fileDataProvider->pagination->pageSize = self::DEFAULT_PAGE_SIZE;
-        
-        // Directory data provider setup
         $directoryDataProvider = new ActiveDataProvider([
             'query' => StorageDirectory::find()
                 ->andWhere(['id_parent' => $id_directory])
@@ -39,9 +38,7 @@ class DefaultController extends Controller
                 'pageSize' => self::DEFAULT_PAGE_SIZE-1,
             ],
         ]);
-        
         $isPicker = Yii::$app->request->get('isPicker', false);
-        
         if (Yii::$app->request->isPjax) {
             return $this->renderAjax('_item-list', [
                 'directoryDataProvider' => $directoryDataProvider,
@@ -49,7 +46,6 @@ class DefaultController extends Controller
                 'isPicker' => $isPicker,
             ]);
         }
-        
         return $this->render('index', [
             'model' => $model,
             'dataProvider' => $dataProvider,
@@ -61,6 +57,9 @@ class DefaultController extends Controller
 
     public function actionUploadFile()
     {
+        if (!\Yii::$app->user->can('storageWebDefaultUploadFile') && !\Yii::$app->user->can('storageWebDefaultUploadFileOwn')) {
+            throw new \yii\web\ForbiddenHttpException(Module::t('You are not allowed to access this page.'));
+        }
         $post = Yii::$app->request->post();
         $type = $post['Storage']['type'] ?? 'file';
         $model = ($type === 'folder') ? new StorageDirectory() : new Storage();
@@ -116,6 +115,9 @@ class DefaultController extends Controller
 
     public function actionDownloadFile()
     {
+        if (!\Yii::$app->user->can('storageWebDefaultDownloadFile') && !\Yii::$app->user->can('storageWebDefaultDownloadFileOwn')) {
+            throw new \yii\web\ForbiddenHttpException(Module::t('You are not allowed to access this page.'));
+        }
         $id = Yii::$app->request->post('id');
         $file = Storage::findOne($id);
 
@@ -134,21 +136,21 @@ class DefaultController extends Controller
 
     public function actionRenameFile($id)
     {
+        if (!\Yii::$app->user->can('storageWebDefaultRenameFile') && !\Yii::$app->user->can('storageWebDefaultRenameFileOwn')) {
+            throw new \yii\web\ForbiddenHttpException(Module::t('You are not allowed to access this page.'));
+        }
         $model = Storage::findOne($id);
         if (!$model) {
             Yii::$app->session->setFlash('error', Module::t('File not found!'));
             return '';
         }
-
         $storagePath = Yii::getAlias('@app') . '/../' . Yii::$app->setting->getValue('storage::path');
         $filePath = $storagePath . '/' . $model->name;
-
         if (!file_exists($filePath)) {
             Storage::deleteAll(['id_storage' => $model->id_storage]);
             Yii::$app->session->setFlash('error', Module::t('File not found!'));
             return '';
         }
-
         if (Yii::$app->request->isPost) {
             if ($model->load(Yii::$app->request->post()) && $model->save())
                 Yii::$app->session->setFlash('success', Module::t('File renamed successfully!'));
@@ -161,6 +163,9 @@ class DefaultController extends Controller
 
     public function actionUpdateFile($id)
     {
+        if (!\Yii::$app->user->can('storageWebDefaultUpdateFile') && !\Yii::$app->user->can('storageWebDefaultUpdateFileOwn')) {
+            throw new \yii\web\ForbiddenHttpException(Module::t('You are not allowed to access this page.'));
+        }
         $model = Storage::findOne($id);
         if (!$model) {
             Yii::$app->session->setFlash('error', Module::t('File not found!'));
@@ -213,6 +218,9 @@ class DefaultController extends Controller
 
     public function actionShareFile($id)
     {
+        if (!\Yii::$app->user->can('storageWebDefaultShareFile') && !\Yii::$app->user->can('storageWebDefaultShareFileOwn')) {
+            throw new \yii\web\ForbiddenHttpException(Module::t('You are not allowed to access this page.'));
+        }
         $model = Storage::findOne($id);
         return $this->renderPartial('_share', [
             'model' => $model,
@@ -221,6 +229,9 @@ class DefaultController extends Controller
 
     public function actionCopyFile()
     {
+        if (!\Yii::$app->user->can('storageWebDefaultCopyFile') && !\Yii::$app->user->can('storageWebDefaultCopyFileOwn')) {
+            throw new \yii\web\ForbiddenHttpException(Module::t('You are not allowed to access this page.'));
+        }
         $id = Yii::$app->request->post('id');
         $sourceModel = Storage::findOne($id);
         if (!$sourceModel) {
@@ -250,6 +261,9 @@ class DefaultController extends Controller
 
     public function actionDeleteFile()
     {
+        if (!\Yii::$app->user->can('storageWebDefaultDeleteFile') && !\Yii::$app->user->can('storageWebDefaultDeleteFileOwn')) {
+            throw new \yii\web\ForbiddenHttpException(Module::t('You are not allowed to access this page.'));
+        }
         $fileId = Yii::$app->request->post('id');
 
         if (Yii::$app->request->isPost && Yii::$app->request->validateCsrfToken()) {
@@ -273,6 +287,9 @@ class DefaultController extends Controller
 
     public function actionPickerModal()
     {
+        if (!\Yii::$app->user->can('storageWebDefaultPickerModal') && !\Yii::$app->user->can('storageWebDefaultPickerModalOwn')) {
+            throw new \yii\web\ForbiddenHttpException(Module::t('You are not allowed to access this page.'));
+        }
         $query = Storage::find();
         $searchModel = new StorageSearch();
         $id_directory = Yii::$app->request->get('id_directory');
@@ -334,6 +351,9 @@ class DefaultController extends Controller
 
     public function actionFileList()
     {
+        if (!\Yii::$app->user->can('storageWebDefaultFileList') && !\Yii::$app->user->can('storageWebDefaultFileListOwn')) {
+            throw new \yii\web\ForbiddenHttpException(Module::t('You are not allowed to access this page.'));
+        }
         $query = Storage::find();
         $dataProvider = new \portalium\data\ActiveDataProvider([
             'query' => $query,
@@ -349,6 +369,9 @@ class DefaultController extends Controller
 
     public function actionSearch()
     {
+        if (!\Yii::$app->user->can('storageWebDefaultSearch') && !\Yii::$app->user->can('storageWebDefaultSearchOwn')) {
+            throw new \yii\web\ForbiddenHttpException(Module::t('You are not allowed to access this page.'));
+        }
         Yii::$app->response->format = \yii\web\Response::FORMAT_HTML;
 
         $q = Yii::$app->request->get('q', '');
@@ -397,6 +420,9 @@ class DefaultController extends Controller
 
     public function actionNewFolder()
     {
+        if (!\Yii::$app->user->can('storageWebDefaultNewFolder') && !\Yii::$app->user->can('storageWebDefaultNewFolderOwn')) {
+            throw new \yii\web\ForbiddenHttpException(Module::t('You are not allowed to access this page.'));
+        }
         $model = new StorageDirectory();
 
         if (Yii::$app->request->isPost) {
@@ -437,6 +463,9 @@ class DefaultController extends Controller
     
     public function actionRenameFolder($id, $id_directory)
     {
+        if (!\Yii::$app->user->can('storageWebDefaultRenameFolder') && !\Yii::$app->user->can('storageWebDefaultRenameFolderOwn')) {
+            throw new \yii\web\ForbiddenHttpException(Module::t('You are not allowed to access this page.'));
+        }
         $model = StorageDirectory::findOne(['id_directory' => $id]);
         if (!$model) {
             Yii::$app->session->setFlash('error', Module::t('Folder not found!'));
@@ -476,6 +505,9 @@ class DefaultController extends Controller
 
     public function actionDeleteFolder($id, $id_directory = null)
     {
+        if (!\Yii::$app->user->can('storageWebDefaultDeleteFolder') && !\Yii::$app->user->can('storageWebDefaultDeleteFolderOwn')) {
+            throw new \yii\web\ForbiddenHttpException(Module::t('You are not allowed to access this page.'));
+        }
         Yii::$app->response->format = \yii\web\Response::FORMAT_JSON;
 
         $folder = StorageDirectory::findOne(['id_directory' => $id]);
@@ -492,6 +524,9 @@ class DefaultController extends Controller
 
     protected function deleteFolderRecursive($folder)
     {
+        if (!\Yii::$app->user->can('storageWebDefaultdeleteFolderRecursive') && !\Yii::$app->user->can('storageWebDefaultdeleteFolderRecursiveOwn')) {
+            throw new \yii\web\ForbiddenHttpException(Module::t('You are not allowed to access this page.'));
+        }
         $subFolders = StorageDirectory::findAll(['id_parent' => $folder->id_directory]);
         foreach ($subFolders as $subFolder) {
             $this->deleteFolderRecursive($subFolder);
